requirements.txt
httpx
requests
pyyaml
lxml
click
rich
m3u8
sortedcontainers
beautifulsoup4
pywidevine